Move support class to a top-level admin class
