+++ includes/classes/general.php
@@ -20,11 +20,18 @@ class General extends Base {
 	protected $api;
 
 	/**
-	 * GatherContent\Importer\Admin instance
+	 * GatherContent\Importer\Admin\Admin instance
 	 *
-	 * @var GatherContent\Importer\Admin
+	 * @var GatherContent\Importer\Admin\Admin
 	 */
 	protected $admin;
+
+	/**
+	 * GatherContent\Importer\Admin\Support instance
+	 *
+	 * @var GatherContent\Importer\Admin\Support
+	 */
+	protected $support;
 
 	/**
 	 * GatherContent\Importer\importer Sync\Pull instance
@@ -82,6 +89,7 @@ class General extends Base {
 
 		$this->api   = new API( _wp_http_get_object() );
 		$this->admin = new Admin\Admin( $this->api );
+		$this->support = new Admin\Support();
 		$this->debug = new Debug( $this->admin );
 		$this->pull  = new Sync\Pull( $this->api );
 		$this->push  = new Sync\Push( $this->api );
@@ -108,6 +116,7 @@ class General extends Base {
 	 */
 	public function init_hooks() {
 		$this->admin->init_hooks();
+		$this->support->init_hooks();
 		$this->pull->init_hooks();
 		$this->push->init_hooks();
 		$this->ajax_handler->init_hooks();
